Add compatibility with Kotlin

+++ canoe/src/main/java/br/com/sailboat/canoe/helper/EntityHelper.java
@@ -2,6 +2,6 @@ package br.com.sailboat.canoe.helper;
 
 public class EntityHelper {
 
-    public static final int NO_ID = -1;
+    public static final long NO_ID = -1;
 
 }
